+++ fugitive-darkness-models/src/main/resources/db/migration/V1_7_0__add_column_is_unpacking_in_git_repo.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "git_repositories"
+    ADD COLUMN is_unpacking BOOLEAN NOT NULL DEFAULT FALSE;