+++ fugitive-darkness-provider/fugitive-darkness-provider-git/src/main/java/com/zer0s2m/fugitivedarkness/provider/git/ContainerInfoSearchFileMatcherGroupGitRepo.java
@@ -1,0 +1,12 @@
+package com.zer0s2m.fugitivedarkness.provider.git;
+
+public record ContainerInfoSearchFileMatcherGroupGitRepo(
+
+        String group,
+
+        int start,
+
+        int end
+
+) {
+}