+++ fugitive-darkness-provider/fugitive-darkness-provider-git/src/main/java/com/zer0s2m/fugitivedarkness/provider/git/impl/GitRepoSearchJGitCallable.java
@@ -10,13 +10,13 @@ import java.util.concurrent.Callable;
 /**
  * Call the repository search and return the result.
  */
-class GitRepoSearchCallable implements Callable<List<ContainerInfoSearchGitRepo>> {
+class GitRepoSearchJGitCallable implements Callable<List<ContainerInfoSearchGitRepo>> {
 
     private final GitRepoFilterSearch filterSearch;
 
     private final GitRepoManager gitRepoService = new GitRepoManagerImpl();
 
-    public GitRepoSearchCallable(GitRepoFilterSearch filterSearch) {
+    public GitRepoSearchJGitCallable(GitRepoFilterSearch filterSearch) {
         this.filterSearch = filterSearch;
     }
 
@@ -27,7 +27,7 @@ class GitRepoSearchCallable implements Callable<List<ContainerInfoSearchGitRepo>
      */
     @Override
     public List<ContainerInfoSearchGitRepo> call() {
-        return gitRepoService.searchByGrep(filterSearch);
+        return gitRepoService.searchByGrep_jgit(filterSearch);
     }
 
 }